Make my code not so bad.
- Replace all indexing operations with size_t.
- Added -Werror flag to compiler.

+++ c/LinkedList/doubly_linked_list.h
@@ -1,3 +1,5 @@
+#include <stddef.h>
+
 typedef struct Node Node;
 
 struct Node {
@@ -7,12 +9,12 @@ struct Node {
 };
 
 typedef struct LinkedList {
-	int elems;
+	size_t elems;
 	Node *head;
 	Node *tail;
 } LinkedList;
 
 Node *node_new(int val);
 void linked_list_add(LinkedList *list, int val);
-int linked_list_remove(LinkedList *list, int index);
-int linked_list_get(LinkedList *list, int index);
+int linked_list_remove(LinkedList *list, size_t index);
+int linked_list_get(LinkedList *list, size_t index);
